+++ NEWS.md
@@ -1,5 +1,7 @@
 # gtsummary 1.2.6
 
+* Update add_p() custom p-value description to NOT require double escape characters for quotes (#361)
+
 * Bug fix for random effects regression model where coefficients were not exponentiated when requested.  Using `broom.mixed::tidy()` rather than `broom::tidy()` resolved issue.
 
 # gtsummary 1.2.5